requirements.txt
flask
scikit-learn
pandas
numpy
matplotlib
openai
groq
python-dotenv
shap
joblib
